ct/reconciler: Factor out test utils
Pulls out the fake_source and unreliable io and metastore into a test
helper file. It'll be useful to have these separate for the metrics
tests.
